+++ packages/examples/browser-example.ts
@@ -1,9 +1,9 @@
 import 'dotenv/config'
-import { RingApi } from '../ring-client-api'
-import { promisify } from 'util'
-const fs = require('fs'),
-  path = require('path'),
-  express = require('express') // note, express is not installed by default so you will need to run `npm i express`
+import { RingApi } from 'ring-client-api'
+import fs from 'node:fs/promises'
+import path from 'node:path'
+// @ts-expect-error express is not installed by default so you will need to run `npm i express`
+import express from 'express'
 
 /**
  * This example creates an hls stream which is viewable in a browser
@@ -24,6 +24,7 @@ async function example() {
   }
 
   const app = express(),
+    __dirname = new URL('.', import.meta.url).pathname,
     publicOutputDirectory = path.join(__dirname, 'public/output')
 
   app.use('/', express.static(path.join(__dirname, 'public')))
@@ -33,8 +34,8 @@ async function example() {
     )
   })
 
-  if (!(await promisify(fs.exists)(publicOutputDirectory))) {
-    await promisify(fs.mkdir)(publicOutputDirectory)
+  if (!(await fs.stat(publicOutputDirectory))) {
+    await fs.mkdir(publicOutputDirectory)
   }
 
   const call = await camera.streamVideo({