+++ packages/homebridge-ring/camera-source.ts
@@ -1,26 +1,28 @@
-import { RingCamera } from 'ring-client-api'
-import { hap } from './hap'
+import type { RingCamera } from 'ring-client-api'
+import { hap } from './hap.ts'
+import type { SrtpOptions } from '@homebridge/camera-utils'
 import {
   doesFfmpegSupportCodec,
   generateSrtpOptions,
   ReturnAudioTranscoder,
   RtpSplitter,
-  SrtpOptions,
 } from '@homebridge/camera-utils'
-import {
-  AudioStreamingCodecType,
-  AudioStreamingSamplerate,
+import type {
   CameraStreamingDelegate,
-  H264Level,
-  H264Profile,
   PrepareStreamCallback,
   PrepareStreamRequest,
   SnapshotRequest,
   SnapshotRequestCallback,
-  SRTPCryptoSuites,
   StartStreamRequest,
   StreamingRequest,
   StreamRequestCallback,
+} from 'homebridge'
+import {
+  AudioStreamingCodecType,
+  AudioStreamingSamplerate,
+  H264Level,
+  H264Profile,
+  SRTPCryptoSuites,
 } from 'homebridge'
 import { logDebug, logError, logInfo } from 'ring-client-api/util'
 import { debounceTime, delay, take } from 'rxjs/operators'
@@ -36,11 +38,14 @@ import {
   SrtcpSession,
 } from 'werift'
 import type { StreamingSession } from 'ring-client-api/streaming/streaming-session'
-import { OpusRepacketizer } from './opus-repacketizer'
+import { OpusRepacketizer } from './opus-repacketizer.ts'
+import path from 'node:path'
 
-const readFileAsync = promisify(readFile),
-  cameraOfflinePath = require.resolve('../media/camera-offline.jpg'),
-  snapshotsBlockedPath = require.resolve('../media/snapshots-blocked.jpg')
+const __dirname = new URL('.', import.meta.url).pathname,
+  mediaDirectory = path.join(__dirname.replace('/lib', ''), 'media'),
+  readFileAsync = promisify(readFile),
+  cameraOfflinePath = path.join(mediaDirectory, 'camera-offline.jpg'),
+  snapshotsBlockedPath = path.join(mediaDirectory, 'snapshots-blocked.jpg')
 
 function getDurationSeconds(start: number) {
   return (Date.now() - start) / 1000
@@ -326,7 +331,7 @@ class StreamingSessionWrapper {
         try {
           const rtp: RtpPacket | undefined = RtpPacket.deSerialize(message)
           this.streamingSession.sendAudioPacket(rtp)
-        } catch (_) {
+        } catch {
           // deSerialize will sometimes fail, but the errors can be ignored
         }
 
@@ -459,7 +464,7 @@ export class CameraSource implements CameraStreamingDelegate {
 
     try {
       await this.previousLoadSnapshotPromise
-    } catch (_) {
+    } catch {
       // ignore errors
     } finally {
       // clear so another request can be made